functions_py/atm_iv/requirements.txt
Flask
google-cloud-bigquery
pandas
